Adds classes to appease the dart analyzer
Adds codegened libraries

This is a result of running tool/gen_apis.dart on the mildly doctored idl/json files and making some small fixes to hush up the analyzer.

Updates versioning

Modifies generated code to pass dartanalyzer

This is the first change I've made to genereated code and will need to apply this to all subsequent rerolls. Hopefully we'll soon have a better way of taking care of this.